Link to annotations support announcement
Previously, it was unclear what kind of changes to the user experience would result from adding JUnit XML artifacts to the CI build. These changes add a link to the blog post announcing support for GitHub Annotations added based on JUnit XML artifacts, which is the main way by which JUnit XML artifact output is presented to the user.